Implement storing of the host z layer
When the host z layer was not stored, a considerable amount of time
could be spent sequentially searching the vertical grid for the host
z layer before any interpolation can take place. Making Particle
objects store a record of their host z layer enables the use of more
intelligent searching algorithms that in turn leads to a significant
reduction in run times.