Remove squeeze call during input data reads
The old implementation introduced slow downs during FVCOM simulations
due to unnecessary data copies. This update reverts back to the
approach where input data arrays are returned as is, and without any
squeezing of unitary dimensions. After allowing for this in
GOTMDataReader, the new approach appears to work well with both
FVCOM and GOTM input files.