Add separate flag to indicate boundary crossings
Have FVCOM local host element search algorithm return a separate
flag that indicates if a particle has crossed an open or land
boundary in addition to the new host element. If the particle
has crossed a boundary, the last valid element the particle
passed through is returned instead. The purpose of this is to
enable an improved treatment of boundary conditions.