Implement check on FVCOM host element searching
Previously, the local search algorithm used for host element searching
when working with FVCOM outputs did not guarentee finding the host
element. It could, occasionally, faily near to land points - i.e.
it could flag a particle as having left the domain when in-fact it
hasn't. The check implemented here corrects this, and guarentees that
any call to find_host() will correctly return either the new host
element or a flag to say the particle has left the domain. In the
case of the latter, find_host() now guarentees that the host returned
will be the last host the particle passed through before exiting the
domain.