Add None checks for horiz bdy calculator
If the model had been misconfigured and it was possible for a particle
to cross a horizontal boundary but no horizontal boundary condition
calculator had been specified the model would have generated a seg
fault. This update introduces a None check that raises an exception
instead, which is far easier to debug.