Move HBC calculator check outside of while loops
If a land boundary has been crossed, it is only necessary to check
once that a HorizontalBoundaryConditionCalculator -- which knows how
to deal with the crossing -- has been selected. Previously, this
check was performed inside a while loop. This update moves it the
check outside of the while loop, meaning it should only be performed
once during the update.