Remove HBC calculations from mock objects
The application of horizontal boundary conditions in mock objects
was only a temporary solution. Several test cases don't need these
since they are for unbounded domains. The correct way to do this
for cases where horizontal boundary conditions should be applied
is to implement find_host() and get_boundary_intersection()
within the test data reader, then use a boundary condition
calculator, as is done with FVCOM.