Add function stubs
Following the update which added NotImplementedError exceptions to all
DataReader functions, not all of the classes that inherit from DataReader were
updated. Basically, all derived classes must now implement all functions
specified in DataReader. This update ensures that this is the case for
a number of derived classes within mock.py.