Fix bug in FVCOM find boundary intersection
The function was returning coordinates for the line and intersection
point of the first element side it tested, rather than the side
that borders the land. Thus, if a particle pathline traversed
two sides before exiting the domain, if could be reflected from
the wrong one, depending on the order in which the lines were
checked. The dependency on the ordering is why this bug got past
the existing unit tests.