Update vertical boundary condition calculator
Modify set_vertical_grid_vars to return a flag identifying
if a particle has been found in the vertical grid. Only if
this returns an error is the vertical boundary condition
calculator invoked. It now takes a data_reader object, the
time and a particle struct as arguments to its core apply
function. It applies the boundary condition, and calls
update_vertical_grid_vars once more to set k_layers etc. If
this fails, an epsilon is also applied. All uses of epsilon
in FVCOMDataReader have been removed.