Add in_domain check when computing output diags
In the previous commit, no check was added to confirm that a
particle was in the domain before calculating bottom depths
and sea surface elevations at the particle's position. This
resulted in nonsense values being calculated and saved to
file which could, by chance, look plausible. This update
forces these variables to be set to standard fill value that
is used throughout the model code.