Fix failing tests due to package dependency updates
Recent updates to netCDF4's num2date, which is basically a wrapper
for the same function in cftime, were causing tests to fail. The
problem arose as num2date switched to returning a cftime datetime
object rather than a python datetime object, with the former
missing functionality on which PyLag relied. To fix this, PyLag
now uses a recent function added to the cftime package called
num2pydate, which is guarenteed to return a python datetime object.

This update also includes a number of fixes to arakawa unit tests,
which started breaking after scipy switched to using a more recent
version of the qhull library. This changed the ordering of elements
in the Delaunay triangulation, which caused the tests to fail. The
tests have now been updated with the correct reference elements, as
generated by the latest release of Scipy.

Conda and setup.py package restrictions have been updated in line
with the new updates.