Make PyLag-tools a PyLag sub-package
PyLag-tools had become a PyLag dependency, as it was needed to build
the docs. As such, the rationale for keeping it as a separate
repository was weakened. This update brings PyLag-tools into the
main PyLag repository as a sub-package called processing. PyLag's
documentation will be fixed up in a second commit.