Switch to computing input date frequencies on the fly
Reinterpret the config parameter `rounding_interval` to be the
number of seconds input datetimes are rounded to, rather than the
amount of time separating input time points. The latter quantity
is now calculated internally. The change makes it possible to
have low frequency input data (e.g. daily or monthly) with more
presice rounding (e.g. to the nearest hour). In particular, this
makes the model easier to use with time averaged data.