Open boundary flags on Arakawa A grids
Open boundaries were not being flagged properly when using stripy
to create a spherical triangulation for a regional grid. To fix
this, the use of stripy and been abandoned in favour of scipy when
creating a triangulation for a regional Arakawa A grid. Stripy is
still used for making global triangulations.