Land sea element mask
Switch to using the land sea element mask to help identify land
crossings. This is in contrast to using the neighbour array to
identify land crossings, which was the previous method. The new
method has several advantages, including: a) Making it faster
to create the grid metrics file on large grids since the flagging
of land elements was a particularly slow process; b) Removing the
need to constantly check whether elements have two land boundaries
when searching for neighbours (now, these elements are simply
masked); c) on Arakawa A grids, shorten the time needed to create
the seed when many particles start on land.