Fix Shepherd interpolation bug for boundary elements
Shepherd interpolation was using invalid array points (with default
values or uninitialised data) for boundary elements. Inverse
distance weighting would have rendered these points inconsequential
for the majority of circumstances, which is likely why the bug
has not shown up before. This update fixes the problem by passing
in another array which flags which elements should be used and
which should not.