Controlling runs with and without random movements
Redefine the way horizontal and vertical diffusivities are specified.
In the config, include two new parameters for the horizontal and
vertical diffusivities: one controlling the method for providing
eddy diffusivities (currently none of from file) and a second
giving the name of the variable if it is present. Develop this
update for FVCOM and ROMS DataReaders. At the same time, take
away responsibility for checking for compatable config options
from the factor methods in numerics.pyx. Rather, shift this into
the data reader classes, which are the objects that control
access to whether diffusivities and whether they are used or not.