Support for a minimal set of FVCOM variables (#113)
* Make obc_file_name optional

* FVCOM grid metrics file dim/var names

Add dictionaries for fvcom dimension and variables names. Clients
may now use these to specify dimension and variable names should
they differ from the FVCOM default names for any reason. If these
arguments are not used, default values are tried.

* FVCOM siglev array

Make it possible to recreate the the FVCOM siglev array from the
siglay array if it is not present in the FVCOM output file.

* FVCOM nbe array

Implement code to recreate the nbe array from the nv array if the
nbe array is not found within the FVCOM data file.

* Fix various issues associated with the previous commits

* FVCOM longitude limits

Add code that automatically detects and converts FVCOM longitudes
on a grid from 0 - 360 to a grid from -180 to 180.

* FVCOM latitude limits

Enable support for FVCOM latitude grids in which latitudes range
from 0.0 to 180.0 rather than -90.0 to 90.0.